feat: private registry auto detect
